backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
sqlalchemy
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
